% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{account}
\alias{account}
\alias{account,Tomba-method}
\title{Account}
\usage{
account(obj)

\S4method{account}{Tomba}(obj)
}
\arguments{
\item{obj}{Tomba class}
}
\value{
A list your account data containing your pricing,first_name,last_name,email,country,requests used on domain search and email finder, and email verifier.
}
\description{
\code{account} Returns information about the current account.
}
\examples{
\dontrun{
client <- Tomba(key="ta_xxxx",secret="ts_xxxx")
result <- account(obj)
}
}
\references{
\url{https://docs.tomba.io/api/account}
}
