% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addin_open}
\alias{addin_open}
\title{Opens the file that is currently open in the source editor in Stencila}
\usage{
addin_open()
}
\description{
Opens the file that is currently open in the source editor in Stencila
}
