% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addin_register}
\alias{addin_register}
\title{Register the function in the source editor that is currently open}
\usage{
addin_register()
}
\description{
Register the function in the source editor that is currently open
}
