% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flythings.R
\name{addObservations}
\alias{addObservations}
\title{addObservations}
\usage{
addObservations(seriesId, observations)
}
\arguments{
\item{seriesId:}{A long value which represents the series to download.}

\item{observations:}{A numeric value.}
}
\value{
The observations of the selected period.
}
\description{
Return a list which contains the value, timestampt and the id of every obsevations between dates.
}
\examples{
addObservations("foo@foo.com", "foo")
}

