% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R
\name{archive_bulk}
\alias{archive_bulk}
\alias{archive_bulk,Tomba-method}
\title{Archive Bulk Operation}
\usage{
archive_bulk(obj, bulk_type, bulk_id)

\S4method{archive_bulk}{Tomba}(obj, bulk_type, bulk_id)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{bulk_type}{Character. The type of bulk operation.}

\item{bulk_id}{Character or integer. The ID of the bulk operation.}
}
\value{
A list confirming the archive.
}
\description{
Archive a bulk operation.
}
