\name{arlequininR-package}
\alias{arlequininR-package}
\alias{arlequininR}
\docType{package}
\title{
Tools to complement Arlequin
}
\description{
Contains functions for reading and writing files formatted for
Excoffier's Arlequin program.
}
\details{
}
}
\author{
Ryan Raaum

Maintainer: Ryan Raaum <code@raaum.org>
}
\references{
}
\keyword{ package }
\seealso{
\code{\link[genotypeData:genotypeData-package]{genotypeData}}
}
\examples{
}
