% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{as.Date.integer}
\alias{as.Date.integer}
\title{Convert YYYYMM to Date}
\usage{
\method{as.Date}{integer}(x, day = "01")
}
\arguments{
\item{x}{\code{integer} vector.}

\item{day}{\code{character} specifying the day of the month.}
}
\value{
vector of \code{Date}s
}
\description{
Convert YYYYMM to Date
}
