% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption.check.R
\name{assumption.check}
\alias{assumption.check}
\title{Check assumptions for anova analysis}
\usage{
assumption.check(data, phase = NULL, type = NULL)
}
\arguments{
\item{data}{data frame}

\item{phase}{study phase. Mandatory: 'baseline', 'induction', 'learning', or 'transfer'.}

\item{type}{assumption type to be checked. Mandatory: 'outliers', 'normality', 'homogeneity'}
}
\value{
assumptions results
}
\description{
In this function you can analize for outliers, normality and homogeneity assumptions
}
\examples{
df<-data.frame('your data')
assumptions(df, phase = 'baseline', type = 'outliers')
assumptions(df,'baseline', 'outliers')

}
