% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{assumptions}
\alias{assumptions}
\title{Check assumptions for anova analysis}
\usage{
assumptions(data, type = NULL)
}
\arguments{
\item{data}{data frame}

\item{type}{assumption type to be checked}
}
\value{
assumptions results
}
\description{
In this function you can analize for outliers, normality and homogeneity assumptions
}
\examples{
df<-data.frame('your data')
assumptions(df);
assumptions(df,type='outliers')

}
