% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{author_finder}
\alias{author_finder}
\alias{author_finder,Tomba-method}
\title{Author Finder}
\usage{
author_finder(obj, url, webhook_url = NULL)

\S4method{author_finder}{Tomba}(obj, url, webhook_url = NULL)
}
\arguments{
\item{obj}{Tomba class}

\item{url}{a character The URL of the article. For example, "https://clearbit.com/blog/company-name-to-domain-api".}

\item{webhook_url}{Character. Webhook URL for async results (optional, default \code{NULL}).}
}
\value{
A list persons data containing the: first_name,last_name,email,score,department,last_updated.
}
\description{
\code{author_finder} This API endpoint generates or retrieves the most likely email address from a blog post url.
}
\examples{
\dontrun{
client <- Tomba(key="ta_xxxx",secret="ts_xxxx")
result <- author_finder(client,url="https://clearbit.com/blog/company-name-to-domain-api")
}
}
\references{
\url{https://docs.tomba.io/api/finder#author-finder}
}
