% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{autocomplete,Tomba-method}
\alias{autocomplete,Tomba-method}
\title{Company Autocomplete is an API that lets you auto-complete company names and retreive logo and domain information.}
\usage{
\S4method{autocomplete}{Tomba}(obj, search)
}
\arguments{
\item{obj}{Tomba class}

\item{search}{name of the company or website.}
}
\value{
autocomplete data
}
\description{
Company Autocomplete is an API that lets you auto-complete company names and retreive logo and domain information.
}
\references{
\url{https://developer.tomba.io/#autocomplete}
}
