% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\docType{data}
\name{AUTOCOMPLETE_PATH}
\alias{AUTOCOMPLETE_PATH}
\title{Autocomplete path}
\format{
An object of class \code{character} of length 1.
}
\usage{
AUTOCOMPLETE_PATH
}
\description{
Autocomplete path
}
\keyword{datasets}
