% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{blob}
\alias{blob}
\title{Get or create a Blob with request_id}
\usage{
blob(request_id)
}
\arguments{
\item{request_id}{request_id or name of blob}
}
\value{
Blob Object
\itemize{
  \item request_id = "#\{request_id\}"
  \item location = "hdfs://hadoopcluster2/projects/planning/dsp_stage/blob/R/#\{request_id\}/"
  \item type = "R"
  \item status = "COMPLETED"
}
}
\description{
Get or create a Blob with request_id
}
\examples{
blob_1 <- blob(11);
blob_1 <- blob("ab");
blob_1 <- blob("Blob_1");
}
