% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{blobDump}
\alias{blobDump}
\title{Dump a blob for a variable_name}
\usage{
blobDump(blob, variable_name, data)
}
\arguments{
\item{blob}{blob object}

\item{variable_name}{variable_name}

\item{data}{data for dump}
}
\value{
None
}
\description{
Dump a blob for a variable_name
}
\examples{
blob1 <- blob("R_blob_1")
data <- "12"
blobDump(blob1, "v_1", data)
}
