% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{blobDumpFile}
\alias{blobDumpFile}
\title{Dump a blob file for a variable_name}
\usage{
blobDumpFile(blob, variable_name, data, type)
}
\arguments{
\item{blob}{blob object}

\item{variable_name}{variable_name}

\item{data}{data for dump}

\item{type}{file type}
}
\value{
None
}
\description{
Dump a blob file for a variable_name
}
\examples{
blob1 <- blob("R_blob_1")
data <- "12"
blobDumpFile(blob1, "v_1", data, "CSV")
}
