% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{blobListAll}
\alias{blobListAll}
\title{list of all the variables for a blob}
\usage{
blobListAll(blob)
}
\arguments{
\item{blob}{blob object}
}
\value{
list of all the variables for a blob
\itemize{
  \item blobs = "["v_1","v_2","v_3"]"
}
}
\description{
list of all the variables for a blob
}
\examples{
blob1 <- blob("R_blob_1")
var_list <- blobListAll(blob1);
}
