% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{blobLoad}
\alias{blobLoad}
\title{load a blob with a variable name}
\usage{
blobLoad(blob, variable_name)
}
\arguments{
\item{blob}{blob object}

\item{variable_name}{variable_name}
}
\value{
object data stored for variable nam for blob object
}
\description{
load a blob with a variable name
}
\examples{
blob1 <- blob("R_blob_1")
data = blobLoad(blob1, "v_1")
}
