% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{blobLoadFile}
\alias{blobLoadFile}
\title{load a blob file with a variable name}
\usage{
blobLoadFile(blob, variable_name, type)
}
\arguments{
\item{blob}{blob object}

\item{variable_name}{variable_name}

\item{type}{file type}
}
\value{
object data stored for variable nam for blob object
}
\description{
load a blob file with a variable name
}
\examples{
blob1 <- blob("R_blob_1")
data = blobLoad(blob1, "v_1", "CSV")
}
