% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R
\name{bulk_progress}
\alias{bulk_progress}
\alias{bulk_progress,Tomba-method}
\title{Bulk Progress}
\usage{
bulk_progress(obj, bulk_type, bulk_id)

\S4method{bulk_progress}{Tomba}(obj, bulk_type, bulk_id)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{bulk_type}{Character. The type of bulk operation.}

\item{bulk_id}{Character or integer. The ID of the bulk operation.}
}
\value{
A list with progress information.
}
\description{
Get the progress of a bulk operation.
}
