% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{client,Tomba-method}
\alias{client,Tomba-method}
\title{Tomba Client}
\usage{
\S4method{client}{Tomba}(obj, path, query = NULL)
}
\arguments{
\item{obj}{Tomba class}

\item{path}{specific path.}

\item{query}{query list}
}
\description{
Tomba Client
}
