% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{client}
\alias{client}
\title{Tomba Client}
\usage{
client(obj, path, query = NULL)
}
\arguments{
\item{obj}{Tomba class}

\item{path}{specific path.}

\item{query}{query list}
}
\description{
Tomba Client
}
