% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba-class.R
\name{client_delete}
\alias{client_delete}
\alias{client_delete,Tomba-method}
\title{Tomba HTTP DELETE Client}
\usage{
client_delete(obj, path)

\S4method{client_delete}{Tomba}(obj, path)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{path}{Character. The API path (relative to the base URL).}
}
\value{
A list with the parsed API response.
}
\description{
Sends a DELETE request to the Tomba API.
}
