% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{combined_find}
\alias{combined_find}
\alias{combined_find,Tomba-method}
\title{Combined Find (Enrichment)}
\usage{
combined_find(obj, email)

\S4method{combined_find}{Tomba}(obj, email)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{email}{Character. The email address to look up.}
}
\value{
A list with combined person and company enrichment data.
}
\description{
Get combined person and company data from an email address.
}
