% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal.R
\name{companies_search}
\alias{companies_search}
\alias{companies_search,Tomba-method}
\title{Companies Search}
\usage{
companies_search(obj, data = list())

\S4method{companies_search}{Tomba}(obj, data = list())
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{data}{A named list of search parameters.}
}
\value{
A list with matching company data.
}
\description{
Search companies using reverse lookup.
}
