% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{company_find}
\alias{company_find}
\alias{company_find,Tomba-method}
\title{Company Find (Enrichment)}
\usage{
company_find(obj, domain)

\S4method{company_find}{Tomba}(obj, domain)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{domain}{Character. The domain name.}
}
\value{
A list with company enrichment data.
}
\description{
Get company data from a domain.
}
