% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\docType{data}
\name{Component}
\alias{Component}
\title{The abstract base class for all Stencila components}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Component
}
\description{
The abstract base class for all Stencila components
}
\section{\code{serve} method}{

Ensures that this component is being served.
A convienience method which simply serves the `host` and
returns this components URL
}
\keyword{datasets}

