% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\docType{data}
\name{Context}
\alias{Context}
\title{A base for context classes to share implementation of
variable packing and unpacking}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Context
}
\description{
A base for context classes to share implementation of
variable packing and unpacking
}
\keyword{datasets}
