% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{count,Tomba-method}
\alias{count,Tomba-method}
\title{Returns total email addresses we have for one domain.}
\usage{
\S4method{count}{Tomba}(obj, domain)
}
\arguments{
\item{obj}{Tomba class}
}
\value{
domain count data
}
\description{
Returns total email addresses we have for one domain.
}
\references{
\url{https://developer.tomba.io/#email-count}
}
