% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19cite.R
\name{covid19cite}
\alias{covid19cite}
\title{COVID-19 Data Hub Data Sources}
\usage{
covid19cite(x, verbose = TRUE)
}
\arguments{
\item{x}{dataset returned by \code{\link{covid19}}.}

\item{verbose}{logical. Print citation? Default \code{TRUE}.}
}
\value{
\code{data.frame} of data sources.
}
\description{
Retrieve citation and metadata of the data sources.
}
\examples{
\dontrun{

# Download data 
x <- covid19("USA")

# Data Sources
s <- covid19cite(x)
View(s)
}

}
