% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R
\name{create_bulk}
\alias{create_bulk}
\alias{create_bulk,Tomba-method}
\title{Create Bulk Operation}
\usage{
create_bulk(obj, bulk_type, data = list())

\S4method{create_bulk}{Tomba}(obj, bulk_type, data = list())
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{bulk_type}{Character. The type of bulk operation.}

\item{data}{A named list with the bulk operation data.}
}
\value{
A list containing the created bulk operation.
}
\description{
Create a new bulk operation.
}
