% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{create_key}
\alias{create_key}
\alias{create_key,Tomba-method}
\title{Create API Key}
\usage{
create_key(obj, data = list())

\S4method{create_key}{Tomba}(obj, data = list())
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{data}{A named list with key data.}
}
\value{
A list containing the created API key.
}
\description{
Create a new API key.
}
