% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads.R
\name{create_lead}
\alias{create_lead}
\alias{create_lead,Tomba-method}
\title{Create Lead}
\usage{
create_lead(obj, data = list())

\S4method{create_lead}{Tomba}(obj, data = list())
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{data}{A named list with lead data.}
}
\value{
A list containing the created lead.
}
\description{
Create a new lead.
}
