% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads-attributes.R
\name{create_lead_attribute}
\alias{create_lead_attribute}
\alias{create_lead_attribute,Tomba-method}
\title{Create Lead Attribute}
\usage{
create_lead_attribute(obj, data = list())

\S4method{create_lead_attribute}{Tomba}(obj, data = list())
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{data}{A named list with attribute data.}
}
\value{
A list containing the created attribute.
}
\description{
Create a new lead attribute.
}
