% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads-lists.R
\name{create_leads_list}
\alias{create_leads_list}
\alias{create_leads_list,Tomba-method}
\title{Create Leads List}
\usage{
create_leads_list(obj, data = list())

\S4method{create_leads_list}{Tomba}(obj, data = list())
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{data}{A named list with list data.}
}
\value{
A list containing the created leads list.
}
\description{
Create a new leads list.
}
