% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbExport}
\alias{dbExport}
\title{Export SQL Query to file}
\usage{
dbExport(conn, sql, file, verbose = TRUE, ...)
}
\arguments{
\item{conn}{a \code{DBIConnection} object, as returned by \code{\link[DBI]{dbConnect}}.}

\item{sql}{the sql query or path to an .sql file.}

\item{file}{output file name.}

\item{verbose}{if \code{TRUE}, print on progress.}

\item{...}{additional arguments passed to \code{\link[data.table]{fwrite}}.}
}
\value{
\code{NULL}
}
\description{
Export SQL Query to file
}
