% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbImportTable}
\alias{dbImportTable}
\title{Import Table}
\usage{
dbImportTable(conn, table, name, index, overwrite, verbose, na.strings, ...)
}
\arguments{
\item{conn}{\code{DBIConnection} object, as returned by \code{\link[DBI]{dbConnect}}.}

\item{table}{\code{data.table}, \code{data.frame}, or path to a data file.}

\item{name}{the table name.}

\item{index}{character vector of column names to use as index.}

\item{overwrite}{if \code{TRUE}, an existing table of the same name will be overwritten. This argument doesn't change behavior if the table does not exist yet.}

\item{verbose}{if \code{TRUE}, print on progress.}

\item{na.strings}{character vector of strings which are to be interpreted as \code{NA} values by \code{\link[data.table]{fread}}.}

\item{...}{additional arguments passed to \code{\link[data.table]{fread}}.}
}
\value{
\code{NULL}
}
\description{
Import Table
}
