% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R
\name{delete_bulk}
\alias{delete_bulk}
\alias{delete_bulk,Tomba-method}
\title{Delete Bulk Operation}
\usage{
delete_bulk(obj, bulk_type, bulk_id)

\S4method{delete_bulk}{Tomba}(obj, bulk_type, bulk_id)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{bulk_type}{Character. The type of bulk operation.}

\item{bulk_id}{Character or integer. The ID of the bulk operation.}
}
\value{
A list confirming deletion.
}
\description{
Delete a bulk operation.
}
