% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{delete_key}
\alias{delete_key}
\alias{delete_key,Tomba-method}
\title{Delete API Key}
\usage{
delete_key(obj, id)

\S4method{delete_key}{Tomba}(obj, id)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{id}{Character. The ID of the API key to delete.}
}
\value{
A list confirming deletion.
}
\description{
Delete an API key by its ID.
}
