% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads.R
\name{delete_lead}
\alias{delete_lead}
\alias{delete_lead,Tomba-method}
\title{Delete Lead}
\usage{
delete_lead(obj, lead_id)

\S4method{delete_lead}{Tomba}(obj, lead_id)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{lead_id}{Character or integer. The ID of the lead to delete.}
}
\value{
A list confirming deletion.
}
\description{
Delete a lead by its ID.
}
