% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads-attributes.R
\name{delete_lead_attribute}
\alias{delete_lead_attribute}
\alias{delete_lead_attribute,Tomba-method}
\title{Delete Lead Attribute}
\usage{
delete_lead_attribute(obj, id)

\S4method{delete_lead_attribute}{Tomba}(obj, id)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{id}{Character. The ID of the attribute to delete.}
}
\value{
A list confirming deletion.
}
\description{
Delete a lead attribute by its ID.
}
