% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads-lists.R
\name{delete_leads_list}
\alias{delete_leads_list}
\alias{delete_leads_list,Tomba-method}
\title{Delete Leads List}
\usage{
delete_leads_list(obj, id)

\S4method{delete_leads_list}{Tomba}(obj, id)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{id}{Character. The ID of the leads list to delete.}
}
\value{
A list confirming deletion.
}
\description{
Delete a leads list by its ID.
}
