% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{domain_search,Tomba-method}
\alias{domain_search,Tomba-method}
\title{Search emails are based on the website You give one domain name and it returns all the email addresses found on the internet.}
\usage{
\S4method{domain_search}{Tomba}(obj, domain)
}
\arguments{
\item{obj}{Tomba class}

\item{domain}{Domain name from which you want to find the email addresses. For example, "stripe.com".}
}
\value{
domain info and emails
}
\description{
Search emails are based on the website You give one domain name and it returns all the email addresses found on the internet.
}
\references{
\url{https://developer.tomba.io/#domain-search}
}
