% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{domain_search}
\alias{domain_search}
\alias{domain_search,Tomba-method}
\title{Domain search}
\usage{
domain_search(obj, domain, enrich_mobile = NULL, webhook_url = NULL)

\S4method{domain_search}{Tomba}(obj, domain, enrich_mobile = NULL, webhook_url = NULL)
}
\arguments{
\item{obj}{Tomba class}

\item{domain}{Domain name from which you want to find the email addresses. For example, "stripe.com".}

\item{enrich_mobile}{Logical. Enrich with mobile phone data (optional, default \code{NULL}).}

\item{webhook_url}{Character. Webhook URL for async results (optional, default \code{NULL}).}
}
\value{
A list domain data containing the organization name,country,social links, and list of emails.
}
\description{
\code{domain_search} Search emails are based on the website You give one domain name and it returns all the email addresses found on the internet.
}
\examples{
\dontrun{
client <- Tomba(key="ta_xxxx",secret="ts_xxxx")
result <- domain_search(obj,domain="stripe")
}
}
\references{
\url{https://docs.tomba.io/api/finder#domain-search}
}
