% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R
\name{download_bulk}
\alias{download_bulk}
\alias{download_bulk,Tomba-method}
\title{Download Bulk Results}
\usage{
download_bulk(obj, bulk_type, bulk_id)

\S4method{download_bulk}{Tomba}(obj, bulk_type, bulk_id)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{bulk_type}{Character. The type of bulk operation.}

\item{bulk_id}{Character or integer. The ID of the bulk operation.}
}
\value{
A list with the download URL or data.
}
\description{
Download the results of a completed bulk operation.
}
