% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{email_finder,Tomba-method}
\alias{email_finder,Tomba-method}
\title{Generates or retrieves the most likely email address from a domain name, a first name and a last name.}
\usage{
\S4method{email_finder}{Tomba}(obj, domain, fname, lname)
}
\arguments{
\item{obj}{Tomba class}

\item{domain}{The domain name of the company, used for emails. For example, "asana.com".}

\item{fname}{The person's first name. It doesn't need to be in lowercase..}

\item{lname}{The person's last name. It doesn't need to be in lowercase..}
}
\value{
persons data
}
\description{
Generates or retrieves the most likely email address from a domain name, a first name and a last name.
}
\references{
\url{https://developer.tomba.io/#email-finder}
}
