% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{email_format}
\alias{email_format}
\alias{email_format,Tomba-method}
\title{Email Format}
\usage{
email_format(obj, domain)

\S4method{email_format}{Tomba}(obj, domain)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{domain}{Character. The domain name.}
}
\value{
A list with the email format information.
}
\description{
Get the email format used by a domain.
}
