% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{email_verifier}
\alias{email_verifier}
\alias{email_verifier,Tomba-method}
\title{Email Verifier}
\usage{
email_verifier(obj, email, webhook_url = NULL)

\S4method{email_verifier}{Tomba}(obj, email, webhook_url = NULL)
}
\arguments{
\item{obj}{Tomba class}

\item{email}{a character email address you want to verify.}

\item{webhook_url}{Character. Webhook URL for async results (optional, default \code{NULL}).}
}
\value{
A list email data containing the: MX records,SMTP server accepts all,SMTP check,deliverability score,status of the email address,status of the verification.
}
\description{
\code{email_verifier} Verify the deliverability of an email address.
}
\examples{
\dontrun{
client <- Tomba(key="ta_xxxx",secret="ts_xxxx")
result <- email_verifier(obj,email="info@tomba.io")
}
}
\references{
\url{https://docs.tomba.io/api/verifier#email-verifier}
}
