% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end.quest.R
\name{end.quest}
\alias{end.quest}
\title{End questionnaires Analysis - end.quest}
\usage{
\method{end}{quest}(data, quest = NULL, tabulation = F)
}
\arguments{
\item{data}{data frame}

\item{quest}{Ending point questionnaires: BDI, STAIR, STAXI, MRS, SRS -> consider upper case.}

\item{tabulation}{Results of the tabulations, the default value is false. The results includes the qualitative result according the spanish sample.}
}
\value{
univariate anova table for: variability in each experimental group. And if it is the case, the results of the tabulations.
}
\description{
Perform a univariate ANOVA on each questionnaire scores. Also, print the tabulation of some questionnaires.
}
\examples{
df<-data.frame('your data')
end.quest(df, quest = 'BDI')
end.quest(df, quest = 'BDI', tabulation = T)
end.quest(df, 'BDI', T)

}
