\name{extract}
\alias{extract}
\alias{extract,genotypeData-method}

\title{Extract Groups, Markers, Monkeys from genotypeData object}

\description{
}
\usage{
}

\arguments{
}

\details{ Blah ... 
}

\value{
  The function returns a \code{genotypeData} with ...
}

\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\keyword{ methods }

