% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{factorEMC}
\alias{factorEMC}
\title{Factor: Expensive minus Cheap}
\usage{
factorEMC(spread, sdret, ret, by, weight = 1, nyse = TRUE)
}
\arguments{
\item{spread}{the effective bid-ask spread}

\item{sdret}{the standard deviation of returns}

\item{ret}{the return at time T+1}

\item{by}{the time index}

\item{weight}{the weight for value-weighted portfolios}

\item{nyse}{logical vector representing whether the stock is listed in NYSE. It is used for the computation of breakpoints.}
}
\value{
Factor returns for each \code{by}-time
}
\description{
Factor: Expensive minus Cheap
}
