% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{factorIML}
\alias{factorIML}
\title{Factor: Illiquid minus Liquid}
\usage{
factorIML(illiq, sdret, ret, by, weight = 1)
}
\arguments{
\item{illiq}{the illiquidity measure}

\item{sdret}{the standard deviation of returns}

\item{ret}{the return at time T+1}

\item{by}{the time index}

\item{weight}{the weight for value-weighted portfolios}
}
\value{
Factor returns for each \code{by}-time
}
\description{
Factor: Illiquid minus Liquid
}
