% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{factorMKT}
\alias{factorMKT}
\title{Factor: Market}
\usage{
factorMKT(ret, rf, by, weight = 1)
}
\arguments{
\item{ret}{the return at time T}

\item{rf}{the risk-free rate at time T}

\item{by}{the time index}

\item{weight}{the weight for the value-weighted market}
}
\value{
Factor returns for each \code{by}-time
}
\description{
Factor: Market
}
