% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-storer.R
\docType{data}
\name{FileStorer}
\alias{FileStorer}
\title{A file storer}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
FileStorer
}
\description{
A file storer
}
\section{new()}{


Create a new \code{FileStorer}

\describe{
  \item{path}{Local file system path. Default \code{''}}
  \item{version}{Version of storer. Currently ignored but required for compatability. Default \code{NULL}}
}
}

\section{readFile()}{


Read content from the file at \code{path}

\describe{
  \item{path}{Path within the storer}
}
}

\section{writeFile()}{


Write content to the file at \code{path}

\describe{
  \item{path}{Path within the storer}
  \item{content}{Content to write}
}
}

\section{listFiles()}{


List files within this storer
}

\section{getInfo()}{


Get information about this storer:

\describe{
  \item{path}{The absolute file sytem path of the storer}
  \item{files}{A list of file in the storer}
}
}

\keyword{datasets}
