% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm.R
\name{fm2fig}
\alias{fm2fig}
\title{Convert Fama-Macbeth regression to ggplot object}
\usage{
fm2fig(x, variable, n = 1, alpha = 0.1, ...)
}
\arguments{
\item{x}{the output of \link{fm}.}

\item{variable}{the name of the column to plot.}

\item{n}{number of periods for rolling averages.}

\item{alpha}{transparency of ribbon.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_ribbon}}}
}
\value{
\code{\link[ggplot2]{ggplot}} object
}
\description{
Convert Fama-Macbeth regression to ggplot object
}
