% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm.R
\name{fm2ggplot}
\alias{fm2ggplot}
\title{Convert Fama-Macbeth regression to ggplot object}
\usage{
fm2ggplot(
  x,
  variable,
  n = 1,
  ci = 0.95,
  fill = "grey",
  alpha = 0.1,
  gg = ggplot(),
  ...
)
}
\arguments{
\item{x}{the output of \link{fm}.}

\item{variable}{the name of the column to plot.}

\item{n}{number of periods for rolling averages.}

\item{ci}{confidence intervals to plot.}

\item{fill}{color of the confidence intervals.}

\item{alpha}{transparency of confidence intervals.}

\item{gg}{a \code{\link[ggplot2]{ggplot}} object to initialize the plot.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_line}} aesthetic mappings.}
}
\value{
\code{\link[ggplot2]{ggplot}} object
}
\description{
Convert Fama-Macbeth regression to ggplot object
}
