% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm.R
\name{fm2tex}
\alias{fm2tex}
\title{Convert Fama-Macbeth regression to TexReg object}
\usage{
fm2tex(x, scale = 1)
}
\arguments{
\item{x}{the output of \link{fm}.}

\item{scale}{multiply the regression coefficients by this value.}
}
\value{
output of \code{\link[texreg]{createTexreg}}
}
\description{
Convert Fama-Macbeth regression to TexReg object
}
