% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm.R
\name{fm2texreg}
\alias{fm2texreg}
\title{Convert Fama-Macbeth regression to TexReg object}
\usage{
fm2texreg(x)
}
\arguments{
\item{x}{the output of \link{fm}.}
}
\value{
output of \code{\link[texreg]{createTexreg}}
}
\description{
Convert Fama-Macbeth regression to TexReg object
}
