% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{function_register}
\alias{function_register}
\title{Register a function}
\usage{
function_register(name_or_path, func, xml)
}
\arguments{
\item{name_or_path}{The name of the function or filesystem path to it's source code}

\item{func}{A function. Optional.}

\item{xml}{A XML specification of the function. Optional.}
}
\description{
Register a function
}
