% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R
\name{get_bulk}
\alias{get_bulk}
\alias{get_bulk,Tomba-method}
\title{Get Bulk Operation}
\usage{
get_bulk(obj, bulk_type, bulk_id)

\S4method{get_bulk}{Tomba}(obj, bulk_type, bulk_id)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{bulk_type}{Character. The type of bulk operation.}

\item{bulk_id}{Character or integer. The ID of the bulk operation.}
}
\value{
A list containing bulk operation details.
}
\description{
Get a specific bulk operation by type and ID.
}
