% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{get_key}
\alias{get_key}
\alias{get_key,Tomba-method}
\title{Get API Key}
\usage{
get_key(obj, id)

\S4method{get_key}{Tomba}(obj, id)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{id}{Character. The ID of the API key.}
}
\value{
A list containing the API key details.
}
\description{
Get a specific API key by its ID.
}
