% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads.R
\name{get_lead}
\alias{get_lead}
\alias{get_lead,Tomba-method}
\title{Get Lead}
\usage{
get_lead(obj, lead_id)

\S4method{get_lead}{Tomba}(obj, lead_id)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{lead_id}{Character or integer. The ID of the lead.}
}
\value{
A list containing lead details.
}
\description{
Get a specific lead by its ID.
}
