% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{get_location}
\alias{get_location}
\alias{get_location,Tomba-method}
\title{Get Location}
\usage{
get_location(obj, domain)

\S4method{get_location}{Tomba}(obj, domain)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{domain}{Character. The domain name.}
}
\value{
A list with the location data for the domain.
}
\description{
Get the location of a domain.
}
