% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.phase.R
\name{graph.phase}
\alias{graph.phase}
\title{graph.phase}
\usage{
graph.phase(
  data,
  phase = NULL,
  quest = NULL,
  score = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL
)
}
\arguments{
\item{data}{data frame}

\item{phase}{analysis phases. Mandatory parameter, it could be: 'baseline' 'induction', 'learning', or 'transfer'.}

\item{quest}{Induction questionnaires: EVEA or SAM -> consider upper case}

\item{score}{Dependent variable you'll use, depending on the questionnaire: anger, sadness, fear, happiness for the EVEA; arousal, valence for the SAM.}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{ylim}{y-axis limit}
}
\value{
specific graphic useful in each experimental phase
}
\description{
specific graphs for each phase
}
\examples{
df<-data.frame('your data')
graph.phase(df,phase='baseline')
graph.phase(df,phase='induction', quest = 'EVEA', score = 'sadness')
graph.phase(df, 'induction', EVEA', sadness')

}
