% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host.R
\docType{data}
\name{host}
\alias{host}
\title{The singleton instance of the \code{Host} class}
\format{An object of class \code{Host} (inherits from \code{R6}) of length 14.}
\usage{
host
}
\description{
The singleton instance of the \code{Host} class
}
\keyword{datasets}
