% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-http-server.R
\docType{data}
\name{HostHttpServer}
\alias{HostHttpServer}
\title{A HTTP server for a Host}
\format{\code{R6Class}.}
\usage{
HostHttpServer
}
\value{
{string} A ticket
}
\description{
Normally there is no need to create a new \code{HostHttpServer}, instead
use the \code{start} method of the \code{host} instance.
}
\section{new()}{


\describe{
  \item{host}{The host to be served}
  \item{address}{The port to listen on. Default \code{'127.0.0.1'}}
  \item{port}{The port to listen on. Default \code{2000}}
  \item{authorization}{Authorization for all requests. Default \code{TRUE}}
}
}

\section{start()}{


Start the server
}

\section{stop()}{


Stop the server
}

\section{handle()}{


Handle a HTTP request
}

\section{route()}{


Route a HTTP request
}

\section{options()}{


Handle a OPTIONS request
}

\section{home()}{


Handle a request to `home`
}

\section{static()}{


Handle a request for a static file
}

\section{manifest()}{


Handle a request for the host's manifest
}

\section{startup()}{


Handle a request to start and environment
}

\section{post()}{


Handle a POST request



Handle a GET request
}

\section{put()}{


Handle a PUT request
}

\section{delete()}{


Handle a DELETE request
}

\section{args()}{


Convert the body into a named list of arguments
}

\section{accepts_json()}{


Does a request accept JSON?
}

\section{ticket_create()}{


Create a ticket (a single-use access token)
}

\section{ticket_check()}{


Check that a ticket is valid. If it is, then it is removed
from the list of tickets and \code{TRUE} is returned. Otherwise,
returns \code{FALSE}
}

\section{ticketed_url()}{

Create a URL with a ticket query parameter so users
can connect to this server
}

\section{token_create()}{


Create a token (a multiple-use access token)
}

\section{token_check()}{


Check that a token is valid.
}

\section{error_403()}{


Generate a 403 error
}

\section{error_404()}{


Generate a 404 error
}

\section{error_500()}{


Generate a 500 error
}

\section{url}{


The URL of the server, or \code{NULL} if not yet started
}

\keyword{datasets}
