% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-http-server.R
\docType{data}
\name{HttpServer}
\alias{HttpServer}
\alias{HostHttpServer}
\title{A HttpServer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HostHttpServer
}
\description{
Normally there is no need to create a new \code{HttpServer}. Instead
use the \code{serve} method of \code{host}
}
\section{\code{initialize} method}{

\describe{
  \item{host}{The host to be served}
  \item{address}{The address to listen on}
  \item{address}{The port to listen on}
}
}

\section{\code{start} method}{


Start the server
}

\section{\code{stop} method}{


Stop the server
}

\keyword{datasets}
