% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induction.effect.R
\name{induction.effect}
\alias{induction.effect}
\title{Induction Effect Anova - induction.effect}
\usage{
induction.effect(
  data,
  quest = NULL,
  score = NULL,
  var.eq = T,
  simple.effect = F
)
}
\arguments{
\item{data}{data frame}

\item{quest}{Induction questionnaires: EVEA or SAM -> consider upper case.}

\item{score}{Dependent variable you'll use, depending on the induction phase questionnaire: anger, sadness, fear, happiness for the EVEA; arousal, valence for the SAM.}

\item{var.eq}{Equality of variance, the default value is true; when false perform whelch's t-test}

\item{simple.effect}{If it is true and if there was a statistical significant result, it will be performed a simple effect analysis of variance in the particular level of time factor.}
}
\value{
anova table for: induction effect at each time point. And if it is the case, the results of the simple effect.
}
\description{
Effect of Induction at each time point -> Only in induction phase. Also, it can perform simple effect analysis in specific time point.
}
\examples{
df<-data.frame('your data')
twoWay.aov(df, quest = 'EVEA', score = 'sadness')
twoWay.aov(df, quest = 'EVEA', score = 'sadness', var.eq = F, simple.effect = T)
twoWay.aov(df, 'EVEA', sadness', simple.effect = T)
twoWay.aov(df, 'EVEA', sadness', F, T)

}
