% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{install}
\alias{install}
\title{Install the Stencila host}
\usage{
install(...)
}
\arguments{
\item{...}{Arguments to pass to \code{host$install}}
}
\description{
Install the Stencila host
}
\seealso{
\code{Host}
}
