% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman.R
\name{kalman}
\alias{kalman}
\title{Kalman filter}
\usage{
kalman(x, nmin, nmax)
}
\arguments{
\item{x}{numeric vector of measurements}

\item{nmin}{the minimum number of observations to estimate the rolling variances}

\item{nmax}{the maximum number of observations to estimate the rolling variances}
}
\value{
numeric vector
}
\description{
This function returns the filtered series from a vector of measurements.
}
\details{
The underlying model is assumed to be:

\deqn{X_t = X_{t-1} + \epsilon_t} 
\deqn{Z_t = X_t + u_t}

where \eqn{X_t} denotes the unobsrved state and \eqn{Z_t} is the corresponding measure.
}
