% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R
\name{launch_bulk}
\alias{launch_bulk}
\alias{launch_bulk,Tomba-method}
\title{Launch Bulk Operation}
\usage{
launch_bulk(obj, bulk_type, bulk_id)

\S4method{launch_bulk}{Tomba}(obj, bulk_type, bulk_id)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{bulk_type}{Character. The type of bulk operation.}

\item{bulk_id}{Character or integer. The ID of the bulk operation.}
}
\value{
A list confirming the launch.
}
\description{
Launch a bulk operation to begin processing.
}
