% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{library_register}
\alias{library_register}
\title{' Register a library of functions
'
Registers the function in each `.R` file in a library directory.}
\usage{
library_register(path)
}
\arguments{
\item{path}{File system path to the library}
}
\description{
' Register a library of functions
'
Registers the function in each `.R` file in a library directory.
}
