% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{linkedin_finder}
\alias{linkedin_finder}
\alias{linkedin_finder,Tomba-method}
\title{Author Finder}
\usage{
linkedin_finder(obj, url, webhook_url = NULL)

\S4method{linkedin_finder}{Tomba}(obj, url, webhook_url = NULL)
}
\arguments{
\item{obj}{Tomba class}

\item{url}{a character The URL of the Linkedin. For example, "https://www.linkedin.com/in/alex-maccaw-ab592978".}

\item{webhook_url}{Character. Webhook URL for async results (optional, default \code{NULL}).}
}
\value{
A list persons data containing the: first_name,last_name,email,score,department,last_updated.
}
\description{
\code{linkedin_finder} This API endpoint generates or retrieves the most likely email address from a Linkedin URL.
}
\examples{
\dontrun{
client <- Tomba(key="ta_xxxx",secret="ts_xxxx")
result <- linkedin_finder(client,url="httpshttps://www.linkedin.com/in/alex-maccaw-ab592978")
}
}
\references{
\url{https://docs.tomba.io/api/finder#linkedin-finder}
}
