% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{list_flags}
\alias{list_flags}
\alias{list_flags,Tomba-method}
\title{List Flags}
\usage{
list_flags(obj, page = NULL, limit = NULL)

\S4method{list_flags}{Tomba}(obj, page = NULL, limit = NULL)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{page}{Integer. Page number for pagination (optional).}

\item{limit}{Integer. Number of results per page (optional).}
}
\value{
A list containing flagged emails.
}
\description{
Get all flagged email addresses.
}
