% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{list_keys}
\alias{list_keys}
\alias{list_keys,Tomba-method}
\title{List API Keys}
\usage{
list_keys(obj)

\S4method{list_keys}{Tomba}(obj)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}
}
\value{
A list containing API keys.
}
\description{
Get all API keys associated with your account.
}
