% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads-attributes.R
\name{list_lead_attributes}
\alias{list_lead_attributes}
\alias{list_lead_attributes,Tomba-method}
\title{List Lead Attributes}
\usage{
list_lead_attributes(obj)

\S4method{list_lead_attributes}{Tomba}(obj)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}
}
\value{
A list containing lead attributes.
}
\description{
Get all lead attributes.
}
