% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads-lists.R
\name{list_leads_lists}
\alias{list_leads_lists}
\alias{list_leads_lists,Tomba-method}
\title{List Leads Lists}
\usage{
list_leads_lists(obj)

\S4method{list_leads_lists}{Tomba}(obj)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}
}
\value{
A list containing leads lists.
}
\description{
Get all leads lists.
}
