% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flythings.R
\name{login}
\alias{login}
\title{login}
\usage{
login(userId, password)
}
\arguments{
\item{userId:}{A numeric value.}

\item{password:}{A numeric value.}
}
\value{
The square of the input.
}
\description{
login to the flythings platform
}
\examples{
login("foo@foo.com", "foo")
}

