% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{logs,Tomba-method}
\alias{logs,Tomba-method}
\title{Returns a your last 1,000 requests you made during the last 3 months.}
\usage{
\S4method{logs}{Tomba}(obj)
}
\arguments{
\item{obj}{Tomba class}
}
\value{
requests data
}
\description{
Returns a your last 1,000 requests you made during the last 3 months.
}
\references{
\url{https://developer.tomba.io/#logs}
}
