% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{logs}
\alias{logs}
\alias{logs,Tomba-method}
\title{Logs}
\usage{
logs(obj, page = NULL, limit = NULL)

\S4method{logs}{Tomba}(obj, page = NULL, limit = NULL)
}
\arguments{
\item{obj}{Tomba class}

\item{page}{Integer. Page number for pagination (optional).}

\item{limit}{Integer. Number of results per page (optional).}
}
\value{
A list requests data containing the: url and The User Agent and IP address associated with the Request,The cost false Free true 1 request,The date,The ISO 3166-1 alpha-2 country code.
}
\description{
\code{logs} Returns a your last 1,000 requests you made during the last 3 months.
}
\examples{
\dontrun{
client <- Tomba(key="ta_xxxx",secret="ts_xxxx")
result <- logs(client)
}
}
\references{
\url{https://docs.tomba.io/api/account#logs}
}
