% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long.data.R
\name{long.data}
\alias{long.data}
\title{long.data}
\usage{
long.data(data, phase = NULL, within = NULL, quest = NULL, score = NULL)
}
\arguments{
\item{data}{data frame}

\item{phase}{analysis phases. Mandatory parameter, it could be: 'baseline' 'induction', 'learning', or 'transfer'.}

\item{within}{Number of within comparissons: 1 or 2}

\item{quest}{Induction questionnaires: EVEA or SAM -> consider upper case}

\item{score}{Dependent variable you'll use, depending on the questionnaire: anger, sadness, fear, happiness for the EVEA; arousal, valence for the SAM.}
}
\value{
long form data frame
}
\description{
transform wide form data into long form. Important to do repeated measure ANOVA analysis
}
