% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.aov.R
\name{main.aov}
\alias{main.aov}
\title{Mixed Main Anova - main.aov}
\usage{
main.aov(data, phase = NULL)
}
\arguments{
\item{data}{data frame}

\item{phase}{analysis phases. Mandatory parameter, it could be: 'baseline' 'induction', 'learning', or 'transfer'.}
}
\value{
anova table
}
\description{
Perform the mixed anova for each experimental phase.
}
\examples{
df<-data.frame('your data')
main.aov(df,phase='induction')
main.aov(df,'induction')

}
