% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mape}
\alias{mape}
\title{Mean absolute percentage error}
\usage{
mape(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector of predicted values.}

\item{y}{a numeric vector of true values.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
\code{numeric} value
}
\description{
Mean absolute percentage error
}
