% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_tempAggr.R
\name{MK.tempAggr}
\alias{MK.tempAggr}
\title{MK test and Sen slope calculator}
\usage{
MK.tempAggr(
  data,
  PW.method = "3PW",
  resolution,
  alpha.mk = 95,
  alpha.cl = 90,
  alpha.xhomo = 90,
  alpha.ak = 95,
  seasonal = FALSE,
  seasons = NULL
)
}
\arguments{
\item{data}{a data.frame with the first column being a POSIXct object with tz = "UTC" and the second column the variable to be analysed}

\item{PW.method}{the PW method used: e.g. one among: 3PW, PW, TFPW.Y, TFPW.WS and VCTFPW. The default is 3PW.}

\item{resolution}{It is taken into account to determine the number of ties; a good guess it is the resolution of the instrument or a little bit higher. This parameters can be determinant for the results but not very sensitive}

\item{alpha.mk}{confidence limit for Mk test in percentage. Default value is 95}

\item{alpha.cl}{confidence limit for the confidence limits of the Sen's slope in percentage. Default value is 90}

\item{alpha.xhomo}{confidence limit for the homogeneity between seasons in percentage. Default value is 90}

\item{alpha.ak}{confidence limit for the first lag autocorrelation in percentage. Default value is 95}

\item{seasonal}{set to TRUE if the analysis needs to be performed over user-defined seasons (default is FALSE)}

\item{seasons}{a vector of the same lenght of the number of records in data.tempAgg used to split the data into seasons. It is used only if seasonal = TRUE.}
}
\value{
P probability for the statistical significance. If 3PW is applied, P = max(P.PW, P.TFPW.Y)

ss statistical significance: alpha \% if the test is ss at the alpha confidence level. Default = 95\%. 0 if the test is not ss at the alpha confidence level; -1 if the test is a TFPW.Y false positive at alpha confidence level; -2 if the test is a PW false positive at alpha confidence level

slope Sen's slope in units/y

UCL upper confidence level in units/y

LCL lower confidence level in units/
}
\description{
`MK_tempAggr` performs the MK test and computes the Sen slope on the given time granularity.
}
\details{
The function implements three prewhitening (PW) methods: PW (Yue et al., 2002) and TFPW.Y (trend free PW, Wang and Swail, 2001) to compute the statistical significance and
VCTFPW (Wang, W., Chen, Y., Becker, S., \& Liu, B. (2015). Variance Correction Prewhitening Method for Trend Detection in Autocorrelated Data. J. Hydrol. Eng., 20(12),
4015033-1-04015033-10. https://doi.org/10.1061/(ASCE)HE.1943-5584.0001234.) to compute the Sen\'s slope.
Only the statistically significant (ss) autocorrelation are taken into account for the prewhitening.
The ss of the trends is taken at 95\% confidence limit. The upper and lower confidence limits are given by the 90\% of the all intervals differences distribution.
The significance level is given by the MK test and has therefore no direct relation to the confidences limits. If seasonal Mann-Kendall is applied,
the yearly trend is assigned only if the results of the seasonal test are homogeneous
}
\examples{

}
\references{
WMO-GAW publication N. 133, annex E, p. 26, MULTMK/PARTMK by C. Libiseller

Gilbert, R. O.: Statistical Methods for Environmental Pollution Monitoring, Van Nostrand Reinhold Company, New York, 1987

Collaud Coen, M., Andrews, E., Bigi, A., Romanens, G., Martucci, G., and Vuilleumier, L.: Effects of the prewhitening method, the time granularity and the time segmentation on the Mann-Kendall trend detection and the associated Sen's slope, Atmos. Meas. Tech., https://doi.org/10.5194/amt-2020-178, 2020.

Collaud Coen, M., Andrews, E., Bigi, A., Romanens, G., Martucci, G., and Vuilleumier, L.: Effects of the prewhitening method, the time granularity and the time segmentation on the Mann-Kendall trend detection and the associated Sen's slope, Atmos. Meas. Tech., https://doi.org/10.5194/amt-2020-178, 2020.
}
\author{
Martine Collaud Coen (martine.collaud@meteoswiss.ch), MeteoSwiss (CH) and Alessandro Bigi (abigi@unimore.it), University of Modena and Reggio Emilia (IT)
}
