% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanprewhite_AR.R
\name{nanprewhite.AR}
\alias{nanprewhite.AR}
\title{Compute AR1 prior to prewhite}
\usage{
nanprewhite.AR(data, alpha.ak = 95)
}
\arguments{
\item{data:}{vector to be analysed}

\item{alpha.ak:}{statistical significance in percentage for the first lag autocorrelation (default is 95)}
}
\value{
ak.lag: fist lag autocorrelation coefficient

data.prewhite: data after removing of the first lag autorcorrelation; if ak.lag is statistically significant, original data otherwise

ak.ss: statistical significance of the first lag autocorrelation: alpha.ak is the statistical significance at the alpha.ak level, zero otherwise
}
\description{
Define the first lag autocorrelation coefficient to prewhite the data as an AR(Kmax) function
}
\examples{

}
\references{
Collaud Coen, M., Andrews, E., Bigi, A., Romanens, G., Martucci, G., and Vuilleumier, L.: Effects of the prewhitening method, the time granularity and the time segmentation on the Mann-Kendall trend detection and the associated Sen's slope, Atmos. Meas. Tech., https://doi.org/10.5194/amt-2020-178, 2020.
}
\author{
Martine Collaud Coen (martine.collaud@meteoswiss.ch), MeteoSwiss (CH) and Alessandro Bigi (abigi@unimore.it), University of Modena and Reggio Emilia (IT)
}
