% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nb_tie.R
\name{Nb.tie}
\alias{Nb.tie}
\title{Classify data to compute ties}
\usage{
Nb.tie(data, resolution)
}
\arguments{
\item{data}{this is the vector of data to be analysed}

\item{resolution}{the measurement resolution, i.e. delta value below which 2 measurements are considered equivalent}
}
\value{
t amount of ties in the data
}
\description{
compute the number of data considered equivalent and treated as ties
}
\references{
Collaud Coen, M., Andrews, E., Bigi, A., Romanens, G., Martucci, G., and Vuilleumier, L.: Effects of the prewhitening method, the time granularity and the time segmentation on the Mann-Kendall trend detection and the associated Sen's slope, Atmos. Meas. Tech., https://doi.org/10.5194/amt-2020-178, 2020.
}
\author{
Martine Collaud Coen (martine.collaud@meteoswiss.ch), MeteoSwiss (CH) and Alessandro Bigi (abigi@unimore.it), University of Modena and Reggio Emilia (IT)
}
