% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntile.R
\name{ntile}
\alias{ntile}
\title{Compute the n-tiles of x based on the breakpoints of y}
\usage{
ntile(x, n, y = x)
}
\arguments{
\item{x}{a numeric vector.}

\item{n}{integer (e.g., \code{10} for deciles) or vector of probabilities (e.g. \code{c(0, 0.3, 0.7, 1)} to split in quantiles 0-30, 30-70, 70-100)}

\item{y}{a numeric vector to compute the breakpoints.}
}
\value{
\code{integer} vector
}
\description{
Compute the n-tiles of x based on the breakpoints of y
}
