\name{opani-internal}
\alias{get.username}
\alias{OPANI_HOME}
\alias{OPANI_TOKEN}
\alias{OPANI_URL}
\alias{opani.cachepath}
\alias{opani.local}
\alias{opani.username}
\alias{cache}
\alias{cache.exists}
\alias{emit}
\alias{get.app}
\alias{get.cloudfile}
\alias{get.filenames}
\alias{get.file}
\alias{get.opani}
\alias{get.s3.file}
\alias{get.values}
\alias{is.opani}
\alias{is.worker}
\alias{logged_in}
\alias{login}
\alias{login_required}
\alias{put.opani}
\alias{put.s3.file}
\alias{run.app.local}
\alias{save.s3.video}
\alias{worker_only}
\alias{tid}
\alias{uid}
\alias{wid}
\alias{odir}
\alias{values}
\alias{opani.appuser}
\alias{opani.appname}
\alias{o.figure.name}
\alias{cachepath.init}
\title{Internal Opani Functions}
\description{
 Internal Opani functions
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{ internal }
