% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_atividade.R
\name{pdf_atividade}
\alias{pdf_atividade}
\title{Formato para entrega de atividades (PDF)}
\usage{
pdf_atividade()
}
\description{
Essa função cria um formato R Markdown para gerar documentos de resolução de
atividades para aulas de laboratório.
}
\examples{
\dontrun{
# ---
# title: "Atividade 1"
# author: "Seu nome e sobrenome"
# output: mschuindt::pdf_atividade
# ---
}
}
