% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{person_find}
\alias{person_find}
\alias{person_find,Tomba-method}
\title{Person Find (Enrichment)}
\usage{
person_find(obj, email)

\S4method{person_find}{Tomba}(obj, email)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{email}{Character. The email address to look up.}
}
\value{
A list with person enrichment data.
}
\description{
Get person data from an email address.
}
