% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{phone_finder}
\alias{phone_finder}
\alias{phone_finder,Tomba-method}
\title{Phone Finder}
\usage{
phone_finder(obj, email, webhook_url = NULL)

\S4method{phone_finder}{Tomba}(obj, email, webhook_url = NULL)
}
\arguments{
\item{obj}{Tomba class}

\item{email}{a character The email address you want to find phone.}

\item{webhook_url}{Character. Webhook URL for async results (optional, default \code{NULL}).}
}
\value{
A list phone data containing the: local_format,intl_format,country_code,line_type,carrier,timezones
}
\description{
\code{phone_finder} Search phone are based on the email You give one email and it returns phone data.
}
\examples{
\dontrun{
client <- Tomba(key="ta_xxxx",secret="ts_xxxx")
result <- phone_finder(client,email="******@zapier.com")
}
}
\references{
\url{https://docs.tomba.io/api/phone#phone-finder}
}
