% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable.R
\name{printbold}
\alias{printbold}
\title{Prints an xtable in bold by}
\usage{
printbold(
  x,
  which = NULL,
  each = c("column", "row"),
  max = TRUE,
  NA.string = "",
  type = c("latex", "html"),
  sanitize.text.function = force,
  sanitize.rownames.function = NULL,
  sanitize.colnames.function = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \link[xtable]{xtable}.}

\item{which}{logical matrix specifying bold cells.}

\item{each}{axis along which to compute the max/min. Possible values for \code{each} are "row" or "column".}

\item{max}{\code{TRUE} for max and \code{FALSE} for min.}

\item{NA.string}{string to be used for missing values in table entries.}

\item{type}{type of table to produce. Possible values for type are "latex" or "html".}

\item{sanitize.text.function}{argument passed to \link[xtable]{print.xtable}.}

\item{sanitize.rownames.function}{argument passed to \link[xtable]{print.xtable}.}

\item{sanitize.colnames.function}{argument passed to \link[xtable]{print.xtable}.}

\item{...}{additional arguments passed to \link[xtable]{print.xtable}.}
}
\description{
If 'which' is given it should be a logical matrix specifying bold cells.
Otherwise: in each column or row with numeric data, the maximum or minimum
value is set bold; 'max' can have entries for each column/row, NA means skip.
}
\note{
The code is adopted from \url{https://gist.github.com/floybix/452201}
}
\examples{
library(xtable)
x <- tail(iris[,1:4])
printbold(xtable(x))
printbold(xtable(x), each = "column", max = c(F, NA, NA, T, NA))
printbold(xtable(x), each = "row", max = FALSE)
printbold(xtable(x), x >= 6.5, type = "html")

}
