% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_3PW.R
\name{prob.3PW}
\alias{prob.3PW}
\title{Applies the 3PW method
Estimate the probability of the MK test with the 3PW method. To have the maximal certainty, P is taken as the maximum of P.PW and P.TFPW.Y. Estimate the statistical significance of the MK test as a function of the given confidence level alpha.mk}
\usage{
prob.3PW(P.PW, P.TFPW.Y, alpha.mk)
}
\arguments{
\item{P.PW:}{probability computed from the PW prewhitened dataset}

\item{P.TFPW.Y:}{probability computed from the TFPW_Y prewhitened dataset}

\item{alpha.mk:}{confidence level in percentage for the MK test.}
}
\value{
P: probability of the MK test according to the 3PW method

ss: statistical significance of the MK test (given the alpha.mk)
}
\description{
Applies the 3PW method
Estimate the probability of the MK test with the 3PW method. To have the maximal certainty, P is taken as the maximum of P.PW and P.TFPW.Y. Estimate the statistical significance of the MK test as a function of the given confidence level alpha.mk
}
\examples{

}
\references{
Collaud Coen, M., Andrews, E., Bigi, A., Romanens, G., Martucci, G., and Vuilleumier, L.: Effects of the prewhitening method, the time granularity and the time segmentation on the Mann-Kendall trend detection and the associated Sen's slope, Atmos. Meas. Tech., https://doi.org/10.5194/amt-2020-178, 2020.
}
