% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-context.R
\docType{data}
\name{RContext}
\alias{RContext}
\title{A R context}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RContext
}
\description{
Implements the Stencila `Context` API.
}
\section{\code{new} method}{


Create a new \code{RContext}

\describe{
  \item{global}{Scope of context is the global environment. Default \code{FALSE}}
  \item{closed}{Context is not nested within the global environment. Default \code{FALSE}}
}
}

\section{\code{run} method}{


Run R code within the context's scope (execute a code "chunk")

\describe{
  \item{code}{R code to be executed}
  \item{options}{Any execution options}
}
}

\section{\code{call} method}{


Run R code within the context's global scope (execute a code "chunk")

\describe{
  \item{code}{R code to be executed}
  \item{inputs}{Name of the variable}
  \item{package}{The data package}
}
}
\keyword{datasets}

