% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readFile}
\alias{readFile}
\title{Read file as character string}
\usage{
readFile(file)
}
\arguments{
\item{file}{the file to read.}
}
\value{
\code{character} vector
}
\description{
Read file as character string
}
