% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_params.R
\name{rebuild_params}
\alias{rebuild_params}
\title{read argument from command line
A Honolulu_White Function}
\usage{
rebuild_params(params)
}
\arguments{
\item{params}{Are you a new guy? Defaults to TRUE.}
}
\description{
This function was written by a person with little experieces in R package generating from Honolulu.
}
\examples{
rebuild_params()
}
\keyword{params}
