% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{register}
\alias{register}
\title{Register the Stencila host}
\usage{
register(...)
}
\arguments{
\item{...}{Arguments to pass to \code{host$register}}
}
\description{
Register the Stencila host
}
\seealso{
\code{Host}
}
