% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_function}
\alias{register_function}
\title{Register a function}
\usage{
register_function(name_or_path, func, xml)
}
\arguments{
\item{name_or_path}{The name of the function or filesystem path to it's source code}
}
\description{
Register a function
}
