% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_library}
\alias{register_library}
\title{Register a library of functions}
\usage{
register_library(path)
}
\arguments{
\item{path}{File system path to the library}
}
\description{
Registers the function in each `.R` file in a library directory.
}
