% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R
\name{rename_bulk}
\alias{rename_bulk}
\alias{rename_bulk,Tomba-method}
\title{Rename Bulk Operation}
\usage{
rename_bulk(obj, bulk_type, bulk_id, name)

\S4method{rename_bulk}{Tomba}(obj, bulk_type, bulk_id, name)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{bulk_type}{Character. The type of bulk operation.}

\item{bulk_id}{Character or integer. The ID of the bulk operation.}

\item{name}{Character. The new name for the bulk operation.}
}
\value{
A list confirming the rename.
}
\description{
Rename a bulk operation.
}
