% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{reset_key}
\alias{reset_key}
\alias{reset_key,Tomba-method}
\title{Reset API Key}
\usage{
reset_key(obj, id)

\S4method{reset_key}{Tomba}(obj, id)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{id}{Character. The ID of the API key to reset.}
}
\value{
A list containing the reset key.
}
\description{
Reset an API key by its ID.
}
\examples{
\dontrun{
cl <- Tomba(key = "ta_xxxx", secret = "ts_xxxx")
result <- reset_key(cl, id = "key-id-here")
}
}
\seealso{
\url{https://docs.tomba.io/api/keys#reset-an-api-key}
}
