% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_test.R
\name{S.test}
\alias{S.test}
\title{Compute the S statistic for the MK test}
\usage{
S.test(data, t.time)
}
\arguments{
\item{data}{it is the vector of observations}

\item{t.time}{it is a matrix with the year, month, date vector as individual variables in each column. The number of rows is the number of obsverations.}
}
\value{
a list containing S and n. S is the double sum on the sign of the difference between data pairs (Si) and n is the number of valid data in each year of the time series
}
\description{
Compute the S statistic for the MK test
}
\examples{

}
\references{
Collaud Coen, M., Andrews, E., Bigi, A., Romanens, G., Martucci, G., and Vuilleumier, L.: Effects of the prewhitening method, the time granularity and the time segmentation on the Mann-Kendall trend detection and the associated Sen's slope, Atmos. Meas. Tech., https://doi.org/10.5194/amt-2020-178, 2020.
}
\author{
Martine Collaud Coen, MeteoSwiss (CH) and alessandro.bigi@unimore.it, University of Modena and Reggio Emilia (IT)
}
