% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar.R
\name{similar_domains}
\alias{similar_domains}
\alias{similar_domains,Tomba-method}
\title{Similar Domains}
\usage{
similar_domains(obj, domain)

\S4method{similar_domains}{Tomba}(obj, domain)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{domain}{Character. The domain name.}
}
\value{
A list with similar domain information.
}
\description{
Retrieve domains similar to the given domain.
}
