% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\docType{data}
\name{SOURCES_PATH}
\alias{SOURCES_PATH}
\title{Email Sources path}
\format{
An object of class \code{character} of length 1.
}
\usage{
SOURCES_PATH
}
\description{
Email Sources path
}
\keyword{datasets}
