\name{Spreads-package}
\alias{Spreads-package}
\alias{Spreads}
\docType{package}
\title{
\packageTitle{Spreads}
}
\description{
\packageDescription{Spreads}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Spreads}
\packageIndices{Spreads}

The package loads Spreads environment to the global environment. Spreads 
environment has functions 'RegisterFunction' and 'RemoveFunction' which
save or remove a serialized function from a local file. That file is 
loaded from Spreads.R project for calling custom R functions from .NET.
}
\author{
\packageAuthor{Spreads}

Maintainer: \packageMaintainer{Spreads}
}
\references{
%%~~ Literature or other references for background information ~~
}
\keyword{ Spreads }
\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
