% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite-context.R
\docType{data}
\name{SqliteContext}
\alias{SqliteContext}
\title{A SQLite context}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
SqliteContext
}
\description{
A SQLite context
}
\section{new()}{


Create a new \code{SqliteContext}

\describe{
  \item{dir}{Working directory. Default \code{NULL}}
}
}

\section{runCode()}{


Run R code within the context's scope

\describe{
  \item{code}{SQL code to be executed}
}
}

\section{callCode()}{


Run R code within a local function scope

\describe{
  \item{code}{SQL code to be executed}
  \item{inputs}{A list with a value pack for each input}
}
}

\keyword{datasets}
