% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite-session.R
\docType{data}
\name{SqliteSession}
\alias{SqliteSession}
\title{A SQLite session}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SqliteSession
}
\description{
A \code{SqliteSession} executes SQL code within an \href{SQLite}{https://sqlite.org} database. In addition to SQL statements, the 
\code{execute()} method support some of the \href{special commands}{https://sqlite.org/cli.html#special_commands_to_sqlite3_dot_commands_} 
of the SQLite command line interface (e.g. \code{tables}). Inputs to the \code{execute} method are handled differently, depending upon 
their type. Primitive data are stored as variables that can be inserted into SQL statements (e.g. \code{SELECT mean(height) FROM data WHERE sex=$\{sex\}}).
Tabular data will be saved as a new database table.
}
\section{\code{db} property}{


Connection to the SQLite database
}

\section{\code{execute()} method}{


Execute code within the session

\describe{
  \item{code}{R code to be executed}
  \item{inputs}{Name of the variable}
}
}
\keyword{datasets}

