% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{start}
\alias{start}
\title{Start serving the Stencila host}
\usage{
start(...)
}
\arguments{
\item{...}{Arguments to pass to \code{host$start}}
}
\description{
Start serving the Stencila host
}
\seealso{
\code{Host}
}
