% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{status,Tomba-method}
\alias{status,Tomba-method}
\title{Returns domain status if is webmail or disposable.}
\usage{
\S4method{status}{Tomba}(obj, domain)
}
\arguments{
\item{obj}{Tomba class}
}
\value{
domain status data
}
\description{
Returns domain status if is webmail or disposable.
}
\references{
\url{https://developer.tomba.io/#domain-status}
}
