% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\docType{data}
\name{STATUS_PATH}
\alias{STATUS_PATH}
\title{Domain status path}
\format{
An object of class \code{character} of length 1.
}
\usage{
STATUS_PATH
}
\description{
Domain status path
}
\keyword{datasets}
