% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STD_normale_var.R
\name{STD.normale.var}
\alias{STD.normale.var}
\title{Calculate the normal standard variable Z}
\usage{
STD.normale.var(data, var.data)
}
\description{
Calculate the normal standard variable Z
}
\examples{

}
\references{
Gilbert 1987
}
\author{
Martine Collaud Coen (martine.collaud@meteoswiss.ch), MeteoSwiss (CH) and Alessandro Bigi (abigi@unimore.it), University of Modena and Reggio Emilia (IT)
}
