% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stop_}
\alias{stop_}
\title{Stop serving the Stencila host}
\usage{
stop_()
}
\description{
Not called stop_ because confusingly masks `stats::stop`,
including withing this package. Not exported to be consistent
with `start()`
}
\seealso{
\code{Host}
}
