% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.statistics.R
\name{summary.statistics}
\alias{summary.statistics}
\title{Summary statistics - summary.statistics}
\usage{
\method{summary}{statistics}(data, phase = NULL, type = NULL)
}
\arguments{
\item{data}{data frame}

\item{phase}{analysis phases. Mandatory parameter, it could be: 'baseline' or 'transfer'.}

\item{type}{Output type. It could be: mean_sd, mean, mean_se}
}
\value{
a tibble with the statistics
}
\description{
it return descriptives for two phases: baseline and induction. You can also input the type to make a summary of the data.
}
\examples{
df<-data.frame('your data')
summary.statistics(df,phase='baseline')
summary.statistics(type = 'mean_sd)
long.data(df, 'induction', '1, 'mean_sd')

}
