% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{technology_check}
\alias{technology_check}
\alias{technology_check,Tomba-method}
\title{Technology Check}
\usage{
technology_check(obj, domain)

\S4method{technology_check}{Tomba}(obj, domain)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{domain}{Character. The domain name.}
}
\value{
A list with technologies used by the domain.
}
\description{
Retrieve the technologies used by a domain.
}
