% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texreg.R
\name{texreg}
\alias{texreg}
\title{Convert regression output to a LaTeX table}
\usage{
texreg(l, file = NULL, tstat = FALSE, scale = 1, ...)
}
\arguments{
\item{l}{a statistical model or a list of statistical models. Lists of models can be specified as \code{l = list(model.1, model.2, ...)}. Different object types can also be mixed.}

\item{file}{using this argument, the resulting table is written to a file rather than to the \code{R} prompt. The file name can be specified as a character string.}

\item{tstat}{if \code{TRUE}, prints t-stats instead of standard errors.}

\item{scale}{multiply the regression coefficients (and standard errors) by this value.}

\item{...}{additional arguments passed to \code{\link[texreg]{texreg}}.}
}
\value{
A character object with a regression table and LaTeX markup. 
The object has an additional \code{"texregTable"} class identifier, 
which causes the object to be formatted nicely on screen when printed.
}
\description{
Convert regression output to a LaTeX table
}
