% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textab.R
\name{textab}
\alias{textab}
\title{Convert a data frame to a LaTeX table}
\usage{
textab(
  x,
  file = "",
  type = c("latex", "html"),
  caption = NULL,
  label = NULL,
  align = NULL,
  digits = NULL,
  display = NULL,
  auto = FALSE,
  bold.which = NULL,
  bold.each = c("column", "row"),
  bold.max = NA,
  NA.string = "",
  sanitize.text.function = force,
  sanitize.rownames.function = NULL,
  sanitize.colnames.function = NULL,
  ...
)
}
\arguments{
\item{x}{a data frame.}

\item{file}{using this argument, the resulting table is written to a file rather than to the R prompt. The file name can be specified as a character string.}

\item{type}{type of table to produce. Possible values for type are "latex" or "html".}

\item{caption}{argument passed to \link[xtable]{xtable}.}

\item{label}{argument passed to \link[xtable]{xtable}.}

\item{align}{argument passed to \link[xtable]{xtable}.}

\item{digits}{argument passed to \link[xtable]{xtable}.}

\item{display}{argument passed to \link[xtable]{xtable}.}

\item{auto}{argument passed to \link[xtable]{xtable}.}

\item{bold.which}{logical matrix specifying bold cells.}

\item{bold.each}{axis along which to compute the max/min. Possible values for \code{each} are "row" or "column".}

\item{bold.max}{whether to boldify maximum values (\code{TRUE}) or minimum values (\code{FALSE}).}

\item{NA.string}{string to be used for missing values in table entries.}

\item{sanitize.text.function}{argument passed to \link[xtable]{print.xtable}.}

\item{sanitize.rownames.function}{argument passed to \link[xtable]{print.xtable}.}

\item{sanitize.colnames.function}{argument passed to \link[xtable]{print.xtable}.}

\item{...}{additional arguments passed to \link[xtable]{print.xtable}.}
}
\description{
Converts a data frame to LaTeX or HTML.
If \code{bold.which} is given it should be a logical matrix specifying bold cells.
Otherwise: in each column or row with numeric data, the maximum or minimum
value is set bold; \code{bold.max} can have entries for each column/row, \code{NA} means skip.
}
\note{
The code is adopted from \url{https://gist.github.com/floybix/452201}
}
\examples{
x <- tail(iris[,1:4])
textab(x)
textab(x, bold.each = "column", bold.max = c(FALSE, NA, NA, TRUE))
textab(x, bold.each = "row", bold.max = FALSE)
textab(x, bold.which = x >= 6.5)

}
