% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\docType{class}
\name{Tomba-class}
\alias{Tomba-class}
\alias{Tomba}
\title{Tomba init}
\description{
An S4 class The Tomba API Constructor
}
\section{Slots}{

\describe{
\item{\code{key}}{A character Tomba API KEY}

\item{\code{secret}}{A character Tomba SECRET KEY}
}}

