% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoWay.aov.R
\name{twoWay.aov}
\alias{twoWay.aov}
\title{Two-Way Anova - twoWay.aov}
\usage{
twoWay.aov(data, phase = NULL, quest = NULL, score = NULL, covariate = NULL)
}
\arguments{
\item{data}{data frame}

\item{phase}{analysis phases. Mandatory parameter, it could be: 'induction', 'learning', or 'transfer'.}

\item{quest}{Induction questionnaires: EVEA or SAM -> consider upper case.}

\item{score}{Dependent variable you'll use, depending on the induction phase questionnaire: anger, sadness, fear, happiness for the EVEA; arousal, valence for the SAM.}

\item{covariate}{Variable name to perform ANCOVA: EDS -> consider upper case.}
}
\value{
anova table
}
\description{
Perform the mixed anova for two levels: Time x Induction, in induction phase.
}
\examples{
df<-data.frame('your data')
twoWay.aov(df, phase='induction', quest = 'EVEA', score = 'sadness')
twoWay.aov(df, phase='induction', quest = 'EVEA', score = 'sadness', covariate = 'EDS')
twoWay.aov(df, 'induction', 'EVEA', sadness')
twoWay.aov(df, 'induction', 'EVEA', sadness', 'EDS')

}
