% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{unpack}
\alias{unpack}
\title{Unpack a package into a R value}
\usage{
unpack(package)
}
\arguments{
\item{package}{The package as a \code{list} or JSON string}
}
\value{
A R value
}
\description{
Unpack a package into a R value
}
