% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads.R
\name{update_lead}
\alias{update_lead}
\alias{update_lead,Tomba-method}
\title{Update Lead}
\usage{
update_lead(obj, lead_id, data = list())

\S4method{update_lead}{Tomba}(obj, lead_id, data = list())
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{lead_id}{Character or integer. The ID of the lead to update.}

\item{data}{A named list with lead data to update.}
}
\value{
A list containing the updated lead.
}
\description{
Update an existing lead by its ID.
}
