% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads-attributes.R
\name{update_lead_attribute}
\alias{update_lead_attribute}
\alias{update_lead_attribute,Tomba-method}
\title{Update Lead Attribute}
\usage{
update_lead_attribute(obj, id, data = list())

\S4method{update_lead_attribute}{Tomba}(obj, id, data = list())
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{id}{Character. The ID of the attribute to update.}

\item{data}{A named list with attribute data to update.}
}
\value{
A list containing the updated attribute.
}
\description{
Update a lead attribute by its ID.
}
