% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leads-lists.R
\name{update_leads_list}
\alias{update_leads_list}
\alias{update_leads_list,Tomba-method}
\title{Update Leads List}
\usage{
update_leads_list(obj, id, data = list())

\S4method{update_leads_list}{Tomba}(obj, id, data = list())
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{id}{Character. The ID of the leads list to update.}

\item{data}{A named list with list data to update.}
}
\value{
A list containing the updated leads list.
}
\description{
Update a leads list by its ID.
}
