% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{usage,Tomba-method}
\alias{usage,Tomba-method}
\title{Returns a your monthly requests.}
\usage{
\S4method{usage}{Tomba}(obj)
}
\arguments{
\item{obj}{Tomba class}
}
\value{
requests data
}
\description{
Returns a your monthly requests.
}
\references{
\url{https://developer.tomba.io/#usage}
}
