% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{usage}
\alias{usage}
\alias{usage,Tomba-method}
\title{Usage}
\usage{
usage(obj)

\S4method{usage}{Tomba}(obj)
}
\arguments{
\item{obj}{Tomba class}
}
\value{
A list requests data containing the: usage of the domain,finder,verifier and source from: Website, Google Sheets add-on,api, browser extension, bulk tasks.
}
\description{
\code{usage} Check your monthly requests.
}
\examples{
\dontrun{
client <- Tomba(key="ta_xxxx",secret="ts_xxxx")
result <- usage(client)
}
}
\references{
\url{https://docs.tomba.io/api/account#usage}
}
