% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize a vector}
\usage{
winsorize(x, pmin = 0.01, pmax = 0.99)
}
\arguments{
\item{x}{a numeric vector.}

\item{pmin}{the lowest percentile.}

\item{pmax}{the upper percentile.}
}
\value{
winsorized \code{numeric} vector
}
\description{
Winsorize a vector
}
