% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{account}
\alias{account}
\title{Returns information about the current account.}
\usage{
account(obj)
}
\arguments{
\item{obj}{Tomba class}
}
\value{
your account data
}
\description{
Returns information about the current account.
}
\references{
\url{https://developer.tomba.io/#account-information}
}
