% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{count}
\alias{count}
\title{Returns total email addresses we have for one domain.}
\usage{
count(obj, domain)
}
\arguments{
\item{obj}{Tomba class}

\item{domain}{Domain name from which you want to find the email addresses. For example, "stripe.com".}
}
\description{
Returns total email addresses we have for one domain.
}
\references{
\url{https://developer.tomba.io/#email-count}
}
