% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbCreateIndex}
\alias{dbCreateIndex}
\title{Create Unique Index}
\usage{
dbCreateIndex(conn, name)
}
\arguments{
\item{conn}{a \code{DBIConnection} object, as returned by \code{\link[DBI]{dbConnect}}.}

\item{name}{the table name.}
}
\value{
\code{NULL}
}
\description{
Create Unique Index
}
