% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbImportTable}
\alias{dbImportTable}
\title{Import Table}
\usage{
dbImportTable(conn, file, overwrite, verbose, ...)
}
\arguments{
\item{conn}{a \code{DBIConnection} object, as returned by \code{\link[DBI]{dbConnect}}.}

\item{file}{the file to import.}

\item{overwrite}{if \code{TRUE}, an existing table of the same name will be overwritten. This argument doesn't change behavior if the table does not exist yet.}

\item{verbose}{if \code{TRUE}, print on progress.}

\item{...}{additional arguments passed to \code{\link[data.table]{fread}}.}
}
\value{
\code{NULL}
}
\description{
Import Table
}
