% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{email_sources}
\alias{email_sources}
\title{Find email address source somewhere on the web.}
\usage{
email_sources(obj, email)
}
\arguments{
\item{obj}{Tomba class}

\item{email}{The email address you want to find sources.}
}
\value{
sources data
}
\description{
Find email address source somewhere on the web.
}
\references{
\url{https://developer.tomba.io/#email-sources}
}
