% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{email_verifier}
\alias{email_verifier}
\title{Verify the deliverability of an email address.}
\usage{
email_verifier(obj, email)
}
\arguments{
\item{obj}{Tomba class}

\item{email}{The email address you want to verify.}
}
\description{
Verify the deliverability of an email address.
}
\references{
\url{https://developer.tomba.io/#email-verifier}
}
