% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm.R
\name{fm2ggplot}
\alias{fm2ggplot}
\title{Convert Fama-Macbeth regression to ggplot object}
\usage{
fm2ggplot(x, y, n = 1, ci = 0.95, fill = "grey", alpha = 0.1, ...)
}
\arguments{
\item{x}{vector of time index (x-axis).}

\item{y}{numeric vector of estimates (y-axis).}

\item{n}{number of periods for rolling averages.}

\item{ci}{level of the confidence intervals.}

\item{fill}{color of the confidence intervals.}

\item{alpha}{transparency of confidence intervals.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_line}} aesthetic mappings.}
}
\value{
\code{\link[ggplot2]{ggplot}} object
}
\description{
Convert Fama-Macbeth regression to ggplot object
}
