% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host.R
\docType{data}
\name{host}
\alias{host}
\title{The host instance}
\format{An object of class \code{Host} (inherits from \code{Component}, \code{R6}) of length 36.}
\usage{
host
}
\description{
The host instance
}
\keyword{datasets}

