% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-http-server.R
\docType{data}
\name{HostHttpServer}
\alias{HostHttpServer}
\title{A HTTP server for a Host}
\format{\code{R6Class}.}
\usage{
HostHttpServer
}
\description{
Normally there is no need to create a new \code{HostHttpServer}, instead
use the \code{start} method of the \code{host} instance.
}
\details{
Note for developers: the general approach taken in developing this class is
to mirror implementations in Node and Python to make it easier to port
code between the languages and ensure consisten implementations.
}
\section{new()}{


\describe{
  \item{host}{The host to be served}
  \item{address}{The port to listen on. Default \code{'127.0.0.1'}}
  \item{port}{The port to listen on. Default \code{2000}}
}
}

\section{start()}{


Start the server
}

\section{stop()}{


Stop the server
}

\section{handle()}{


Handle a HTTP request
}

\section{route()}{


Route a HTTP request
}

\section{static()}{


Handle a request for a static file
}

\section{run()}{


Handle a request to call a host method
}

\section{error()}{


Generate an error response
}

\section{error_400()}{


Generate a 400 error
}

\section{error_401()}{


Generate a 401 error
}

\section{error_403()}{


Generate a 403 error
}

\section{error_404()}{


Generate a 404 error
}

\section{error_500()}{


Generate a 500 error
}

\section{address}{


The address of the server
}

\section{port}{


The port of the server
}

\section{url}{


The URL of the server, or \code{NULL} if not yet started
}

\keyword{datasets}
