% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-http-server.R
\docType{data}
\name{HostHttpServer}
\alias{HostHttpServer}
\title{A HTTP server for a Host}
\format{\code{R6Class}.}
\usage{
HostHttpServer
}
\description{
Normally there is no need to create a new \code{HostHttpServer}, instead
use the \code{start} method of the \code{host} instance.
}
\section{new()}{


\describe{
  \item{host}{The host to be served}
  \item{address}{The port to listen on. Default '127.0.0.1'}
  \item{port}{The port to listen on. Default 2000}
}
}

\section{start()}{


Start the server
}

\section{stop()}{


Stop the server
}

\section{handle()}{


Handle a HTTP request
}

\section{route()}{


Route a HTTP request
}

\section{options()}{


Handle OPTIONS request. Necessary for pre-flighted CORS requests.
}

\section{home()}{


Handle a request to `home`
}

\section{static()}{


Handle a request for a static file
}

\section{post()}{


Handle a POST request



Handle a GET request
}

\section{put()}{


Handle a PUT request
}

\section{delete()}{


Handle a DELETE request
}

\section{url}{


The URL of the server, or \code{NULL} if not yet started
}

\keyword{datasets}
