% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host-http-server.R
\docType{data}
\name{HttpServer}
\alias{HttpServer}
\alias{HostHttpServer}
\title{A HTTP server for a `Host`}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HostHttpServer
}
\description{
Normally there is no need to create a new \code{HostHttpServer}, instead
use the \code{start} method of the \code{host} nstance.
}
\section{\code{initialize} method}{

\describe{
  \item{host}{The host to be served}
  \item{port}{The port to listen on}
}
}

\section{\code{start} method}{


Start the server
}

\section{\code{stop} method}{


Stop the server
}

\section{\code{handle} method}{


Handle a HTTP request
}

\section{\code{route} method}{


Route a HTTP request
}

\section{\code{options} method}{


Handle OPTIONS request
Necessary for preflighted CORS requests (https://developer.mozilla.org/en-US/docs/Web/HTTP/Access_control_CORS#Preflighted_requests)
}

\section{\code{home} method}{


Handle a request to `home`



Handle a request for a static file
}

\keyword{datasets}
