% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-http.R
\docType{data}
\name{HttpServer}
\alias{HttpServer}
\title{A HttpServer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HttpServer
}
\description{
Normally there is no need to create a new \code{HttpServer}. Instead
use the \code{serve} method of \code{host}
}
\section{\code{initialize} method}{

\describe{
  \item{host}{The host to be served}
  \item{address}{The address to listen on}
  \item{address}{The port to listen on}
}
}

\section{\code{serve} method}{

\describe{
  \item{on}{The host to be served}
}
}
\keyword{datasets}

