% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induction.effect.R
\name{induction.effect}
\alias{induction.effect}
\title{Induction Effect Anova - induction.effect}
\usage{
induction.effect(data, quest = NULL, score = NULL, var.eq = T)
}
\arguments{
\item{data}{data frame}

\item{quest}{Induction questionnaires: EVEA or SAM -> consider upper case.}

\item{score}{Dependent variable you'll use, depending on the induction phase questionnaire: anger, sadness, fear, happiness for the EVEA; arousal, valence for the SAM.}

\item{var.eq}{Equality of variance, the default value is true; when false perform whelch's t-test}
}
\value{
anova table
}
\description{
Effect of Induction at each time point -> Only in induction phase.
}
\examples{
df<-data.frame('your data')
twoWay.aov(df, quest = 'EVEA', score = 'sadness')
twoWay.aov(df, quest = 'EVEA', score = 'sadness', var.eq = F)
twoWay.aov(df, 'EVEA', sadness')
twoWay.aov(df, 'EVEA', sadness', F)

}
