% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{logs}
\alias{logs}
\title{Returns a your last 1,000 requests you made during the last 3 months.}
\usage{
logs(obj)
}
\arguments{
\item{obj}{Tomba class}
}
\description{
Returns a your last 1,000 requests you made during the last 3 months.
}
\references{
\url{https://developer.tomba.io/#logs}
}
