% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long.data.R
\name{long.data}
\alias{long.data}
\title{long.data}
\usage{
long.data(data, phase = NULL)
}
\arguments{
\item{data}{data frame}

\item{phase}{analysis phases. Mandatory parameter, it could be: 'baseline' 'induction', 'learning', or 'transfer'.}
}
\value{
long form data frame
}
\description{
transform wide form data into long form. Important to do repeated measure ANOVA analysis
}
\examples{
df<-data.frame('your data')
long.data(df,phase='induction')
long.data(df,'induction')

}
