% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{pack}
\alias{pack}
\title{Pack a R value into a package}
\usage{
pack(value)
}
\arguments{
\item{value}{The R value to be packaged}
}
\value{
A package as an R \code{list}
}
\description{
Pack a R value into a package
}
