% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing.R
\name{pack}
\alias{pack}
\title{Pack an object into a data package}
\usage{
pack(object)
}
\arguments{
\item{object}{The R object to be packaged}
}
\value{
Data package as an R list
}
\description{
Pack an object into a data package
}

