% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone.R
\name{phone_validator}
\alias{phone_validator}
\alias{phone_validator,Tomba-method}
\title{Phone Validator}
\usage{
phone_validator(obj, phone)

\S4method{phone_validator}{Tomba}(obj, phone)
}
\arguments{
\item{obj}{A \code{\link{Tomba}} object.}

\item{phone}{Character. The phone number to validate.}
}
\value{
A list with validation data.
}
\description{
Validate a phone number and retrieve metadata.
}
