% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prewhite.R
\name{prewhite}
\alias{prewhite}
\title{Compute the prewhitened datasets}
\usage{
prewhite(data.ts, column, resolution, alpha.ak = 95)
}
\arguments{
\item{data.ts:}{it is a dataframe with a column named "Time" as POSIXct with timezone UTC and the other columns with observations}

\item{column:}{the number column of the data to be analyzed (must be greater than 2)}

\item{resolution:}{the measurement resolution, i.e. delta value below which 2 measurements are considered equivalent. It is used to compute the number of ties}

\item{alpha.ak:}{statistical significance in percentage for the first lag autocorrelation (default is 95)}
}
\value{
data.PW: a dataframe with 6 columns:
     Time (as POSIXct in UTC)
     PW = PW with the first lag autocorrelation of the data
     PW.cor = PW corrected with 1/(1-ak1)
     TFPW.WS = PW with the first lag autocorrelation of the data after detrending computed from PW data (see Wang & Swail)
     TFPW.Y = method of Yue et al 2002, not 1/1-ak1) correction detrend on original data
     VCTFPW = PW with the first lag autocorrelation of the data after detrending + correction of the PW data for the variance (see Wang 2015)
}
\description{
calculate all the necessary prewhitened datasets to asses the statistical significance and to compute the Sen's slope for each of the prewhitening method, including 3PW
}
\examples{

}
\references{
Collaud Coen, M., Andrews, E., Bigi, A., Romanens, G., Martucci, G., and Vuilleumier, L.: Effects of the prewhitening method, the time granularity and the time segmentation on the Mann-Kendall trend detection and the associated Sen's slope, Atmos. Meas. Tech., https://doi.org/10.5194/amt-2020-178, 2020.
}
\author{
Martine Collaud Coen (martine.collaud@meteoswiss.ch), MeteoSwiss (CH) and Alessandro Bigi (abigi@unimore.it), University of Modena and Reggio Emilia (IT)
}
