% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-context.R
\docType{data}
\name{RContext}
\alias{RContext}
\title{A R context}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RContext
}
\description{
Implements the Stencila `Context` API.
}
\section{\code{new} method}{


Create a new \code{RContext}

\describe{
  \item{global}{Scope of context is the global environment. Default \code{FALSE}}
  \item{closed}{Context is not nested within the global environment. Default \code{FALSE}}
}
}

\section{\code{runCode} method}{


Run R code within the context's scope (i.e. execute a code "chunk")

\describe{
  \item{code}{R code to be executed}
  \item{options}{Any execution options}
}
}

\section{\code{callCode} method}{


Run R code within the context's global scope (i.e. execute a code "chunk")

\describe{
  \item{code}{R code to be executed}
  \item{args}{A list with a data pack for each argument}
}
}

\section{\code{codeDependencies} method}{


Returns an array of all variable names not declared within
the piece of code. This might include global functions and variables used
within the piece of code.

\describe{
  \item{code}{R code}
}
}
\keyword{datasets}

