% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-context.R
\docType{data}
\name{RContext}
\alias{RContext}
\title{A R context}
\format{\code{R6Class}.}
\usage{
RContext
}
\description{
An execution context for R code
}
\details{
In Stencila, a "context" is the thing that executes code for a particular programming language.
This is the context for R.
It implements the \code{Context} API so that it can talk to other parts of the platform,
including contexts for other languages, Documents, and Sheets.
}
\section{new()}{


Create a new \code{RContext}

\describe{
  \item{global}{Scope of context is the global environment. Default \code{FALSE}}
  \item{closed}{Context is not nested within the global environment. Default \code{FALSE}}
}
}

\section{runCode()}{


Run R code within the context's scope

\describe{
  \item{code}{R code to be executed}
  \item{options}{Any execution options}
}
}

\section{callCode()}{


Run R code within a local function scope

\describe{
  \item{code}{R code to be executed}
  \item{args}{A list with a data pack for each argument}
}
}

\section{codeDependencies()}{


Returns an array of all variable names not declared within
the piece of code. This might include global functions and variables used
within the piece of code.

\describe{
  \item{code}{R code}
}
}

\examples{
context <- RContext$new()
context$runCode('my_var <- 42')
context$runCode('my_var')
context$callCode('my_var / 7')
context$callCode('x * 2', list(x=pack(2)))
context$callCode('plot(1,1)')
}
\keyword{datasets}
