% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-context.R
\docType{data}
\name{RContext}
\alias{RContext}
\title{A R context}
\format{\code{R6Class}.}
\usage{
RContext
}
\description{
An execution context for R code
}
\details{
In Stencila, a "context" is the thing that executes code for a particular programming language.
This is the context for R.
It implements the \code{Context} API so that it can talk to other parts of the platform,
including contexts for other languages, Documents, and Sheets.
}
\section{new()}{


Create a new \code{RContext}

\describe{
  \item{global}{Scope of context is the global environment. Default \code{FALSE}}
  \item{closed}{Context is not nested within the global environment. Default \code{FALSE}}
}
}

\section{runCode()}{


Run R code within the context's scope

\describe{
  \item{code}{R code to be executed}
  \item{options}{Any execution options}
}
}

\section{callCode()}{


Run R code within a local function scope

\describe{
  \item{code}{R code to be executed}
  \item{args}{A list with a data pack for each argument}
}
}

\section{codeDependencies()}{


Returns an array of all variable names not declared within
the piece of code. This might include global functions and variables used
within the piece of code.

\describe{
  \item{code}{R code}
}
}

\examples{
context <- RContext$new()

# Assign a variable within the context
context$runCode('my_var <- 42')

# Get the variable as an output value
context$runCode('my_var')

# The variable is NOT available in `callCode`
context$callCode('my_var')$errors[[1]]$message

# Intead, you can pass input values
context$callCode('x * y', list(x=pack(6), y=pack(7)))

# Returned output value can include plots
context$callCode('plot(1,1)')$output
}
\keyword{datasets}
