\name{Spreads-package}
\alias{Spreads-package}
\alias{Spreads}
\docType{package}
\title{
\packageTitle{Spreads}
}
\description{
\packageDescription{Spreads}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Spreads}
\packageIndices{Spreads}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Spreads}

Maintainer: \packageMaintainer{Spreads}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
