% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite-context.R
\docType{data}
\name{SqliteContext}
\alias{SqliteContext}
\title{A SQLite context}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
SqliteContext
}
\description{
A SQLite context
}
\section{new()}{


Create a new \code{SqliteContext}

\describe{
  \item{dir}{Working directory. Default \code{NULL}}
}
}

\section{compile()}{


Analyse SQL code and return the names of inputs, outputs
and the returned value

\describe{
  \item{code}{SQL code to be analysed}
  \item{exprOnly}{Ensure that the code is a select expression?}
}
}

\section{execute()}{


Run SQL code

\describe{
  \item{code}{SQL code to be executed}
  \item{inputs}{A list with a value pack for each input}
  \item{exprOnly}{Ensure that the code is a simple expression?}
}
}

\keyword{datasets}
