% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{status}
\alias{status}
\title{Returns domain status if is webmail or disposable.}
\usage{
status(obj, domain)
}
\arguments{
\item{obj}{Tomba class}

\item{domain}{Domain name from which you want to check. For example, "gmail.com".}
}
\description{
Returns domain status if is webmail or disposable.
}
\references{
\url{https://developer.tomba.io/#domain-status}
}
