% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STD_normale_var.R
\name{STD.normale.var}
\alias{STD.normale.var}
\title{Calculate the normal standard variable Z
see Gilbert 1987}
\usage{
STD.normale.var(data, var.data)
}
\description{
Calculate the normal standard variable Z
see Gilbert 1987
}
\examples{

}
\author{
Martine Collaud Coen, MeteoSwiss (CH) and alessandro.bigi@unimore.it, University of Modena and Reggio Emilia (IT)
}
