% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stop_}
\alias{stop_}
\title{Stop serving the Stencila host}
\usage{
stop_(...)
}
\arguments{
\item{...}{Arguments to pass to \code{host$stop}}
}
\description{
Stop serving the Stencila host
}
\seealso{
\code{Host}
}
