% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing.R
\name{unpack}
\alias{unpack}
\title{Unpack a data package into an object}
\usage{
unpack(package)
}
\arguments{
\item{package}{The data package as a \code{list} or JSON string}
}
\value{
A R object
}
\description{
Unpack a data package into an object
}

