% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomba.R
\name{usage}
\alias{usage}
\title{Returns a your monthly requests.}
\usage{
usage(obj)
}
\arguments{
\item{obj}{Tomba class}
}
\description{
Returns a your monthly requests.
}
\references{
\url{https://developer.tomba.io/#usage}
}
